Moved host IP address validation to fetchers
The fetcher config creation was previously raising an
MKIPAddressLookupError in case the config creation could not gather an
resolve an IP address for one of the configured hosts. This lead to a
complete termination of the config creation.

Instead of during config creation, this validation has to be done right
before contacting the host. This change is exactly doing this.

Now the fetcher will raise an exception in the correct moment like this:

```
MKIPAddressLookupError('Failed to lookup IP address and no explicit IP address configured')
```

What's left todo is to handle this exception correctly. This will be
fixed soon with the general error handling changes.

CMK-5295

Change-Id: I78ea68d01aef90d937edb0d7ee21991ee31ee11c