TimeSeries object includes metadata about how to render itself
Commit deals with the refactoring consequences. As TimeSeries includes
render information, there is no need to deal with multi-types on curves,
once for the TimeSeries once for a list of tuples of render data and
TimesSeries.

The evaluation of an expression thus simplifies its return type to
List[TimeSeries]

CMK-5232

Change-Id: I8b7bbadc77aea54901b5834954952cdd337664c5