Change FetcherHeader internal representation to bytes
The payload is now in bytes.  It makes sense to have the header in bytes
as well.  Moreover, typing is slightly safer as we do not need to
convert numbers to string and string to numbers.  The code is also
slightly simpler.

It should also be easier to maintain and extend as we need at least one
more field for the error handling.

CMK-5450

Change-Id: I2dc68cd82e9db7ffc6a71c1d7e93363735b2ac54