check_table: remove deduplication
Note that the removed code is already different from the code of the latest released version.
In verison 1.6 the 'remove_duplicate' flag did *not* guarantee a unique service description,
it only removed duplicates if one service corresponded to an SNMP- and one to an TCP-plugin.
This distiction, however, can no longer be made in the new API.

Thus we drop duplicate removal *at that point* completely. When writing the core configuration
duplicate services will be dropped, and a warning is generated.

CMK-5393

Change-Id: I5bc6e5137b11c7cb6cf4a94591ababb0051f58b4