API: Localize FileCache serialize and deserialize
This change moves the serialization part of FileCache to the FileCache
class, where the deserialization already was.  On top of the better
maintainability, this change brings a series of general improvements:

- Typing of FileCache is stricter and safer:  It is not possible anymore
to instantiate, for example, PiggyBack with a DefaultAgentFileCache, &c.
- Also regarding typing:  The sources now explicitly instantiate and
return the correct FileCache type.
- The sources now directly instantiate FileCache objects instead of
going through serialization and deserialization.
- The new API is consistent with the other factories (parser,
summarizer, &c.)
- The new API should also let us change the `__init__` of the FileCache
classes, for example by emptying NoCache.__init__().

We still group the global variables in a FileCacheFactory class.  The
class may be removed once the globals are not global anymore.

CMK-5450

Change-Id: I463b392a6efb6c23e1fa5d67297ecf6c6d1963ce