Add test to uncover discovery issues before integration tests
The test_bulk_discovery_start_with_defaults integration tests counts the
discovered services when using the default linux agent outout. In case
we break something, the issue is discovered lately during the time consuming
integration tests.

It is better to uncover these issues before merging it into the public branches,
so we create a test (which is obviously more than a unit test) that performs
the discovery on the default linux agent output and verifies the results.

Change-Id: If9650c4599e815b70d63ab025fe4da1991461fe9