Execute disabling of outdated MKPs before cmk-update-config
The outdated MKPs may affect the update logic in cmk-update-config. To reduce
the effect of outdated MKPs on the builtin Checkmk logic, we now execute the
disabling of MKPs during "omd update" before the Checkmk code is executed for
the first time.

This change also adds the "mkp disable-outdated" command.

CMK-5505

Change-Id: I8fe15ed8c09806f6e9d802be6f5ebe4ac6190a9e