API: Improve Fetcher API
We make the field of the Fetchers public and final.

As it comes from outside anyway, it is not the Fetcher's responsibility
to hide it.  Even when it was private, the values were actually
displayed by `__repr__()`.

The remaining private attributes clarify the intent of the class.

The public attributes make the Fetchers fully black-box testable.

It should also makes sense to the user that a TCP fetcher, for example,
has an `address` and an `encryption_settings` attribute.

This also lets us fix some typing issues.

Now, the Fetchers are also more consistent with args/kwargs
in `__init__()`.

CMK-5450

Change-Id: I20cb90ab9c1276925ca4a3fc2f1432da095679e4