additional_rules: Updated valuespec to handle differently named parameters
The main change is in file file_attributes_utils.py
The parameter names for the parameters minage, maxage, minsize, maxsize are dynamically generated as keys during runtime. The parameter names of the valuespec function additional_rules have been changed to input parameters, so that they can be adjusted for each valuespec.

CMK-5537

Change-Id: I3c78c979ea97dd572d5a92c69a07f67c876597ab