Serialize exceptions with pickle
We were using `repr()`/`eval()` but this naive scheme does not let us
transport our special exceptions (defined in `cmk.utils.exceptions`).
Exceptions that are nested in a different module would not be
deserialized properly either.

We replace `repr()`/`eval()` with `pickle`, which is the prefered
serialization method in Python and does not have the same drawbacks.
This further has the advantage of preserving the traceback.

CMK-5472

Change-Id: I867b5815605919a5c530cc1823afc4131290b4c6