Fix recently broke htpasswd user handling
The GUI supports ad-hoc creation of users that only have
an entry in the etc/htpasswd file. This is basically a feature
that was important in the early days of Checkmk / the GUI where
WATO was not managing users in most environments. We keep supporting
this use case for the moment, but may break with this in the future.

Change-Id: I346fa90c1ae485fc698b83c2c2497da5bbc5beb1