Don't output logging information in automation calls
Remove setup_console_logging in cmk.utils.piggyback. This
leaves the initial NullHandler of cmk.utils.log.logger in
place. This way no logging is set up accidentally for
other modules on import of cmk.utils.piggyback or modules
that import it like e.g. the prominent cmk.base.config.

Finally, use clear_console_logging to disable logging in
automation calls. In this case it lead to unwanted output
in the notification-get-bulks automation call which lead
to a crash of the notifacations page in the GUI. The
logging is disabled after all arguments are successfully
parsed and right before the automation handler is called.

Remove an unnecessary argument to check_mk_local_automation
along the way.

CMK-5517

Change-Id: I2ea4ac0fcb57acf52226ffc2b5b3d4c7f8253445