Fix LDAP connection test
Between 1.6 and 2.0 the logic for instanciating the connections
changed a bit. The current code requires the "type" attribute to
be set and point to the connector type of the connection.

However, the config that was created by the LDAP configuration dialog
did not contain this field which made the LDAP connection test fail
with a crash report.

The cmk-update-config command now cares about fixing all existing
configurations and the config dialog is now adding the attribute when
adding a new connection.

CMK-5597

Change-Id: I60aacac984ae6c8c2ded91748da296346edc5828