Complete layer 3 protocol layer
This should help avoid typing the errors between L2 and L3 that are
possible when FetcherMessage accepts bytes.

This should generally encourage users to stay away from bytes.

Other advantages are encapsulation of the serialization and
deserialization of the different types of raw data and consistency with
the rest of the modules.

Change-Id: I70f88036f1a54b189fa62c6356ba6af42fa8e50b