Add svg icons for the new design
Since some icons require an emblem (a small icon in the lower
right corner of the icon) the methods html.render_icon and
html.icon methods now take an optional argument to add an
emblem. The emblems are stored under the name icon_emblem_*.

Some icons are still missing and will follow in later commits.

CMK-5484

Change-Id: Ifdcc427ee180d7386cc681dbed99571e7b05f938