11368 FIX netapp_api_luns: Report correct total size
The check plugin <tt>netapp_api_luns</tt>, which monitors the usage of LUNs of
Netapp filers, can be configured to not apply any levels to the used storage and
only report the total size of the LUN instead. The size report by the check
plugin in this particular case was wrong by a factor of 1024^2, which is fixed
by this werk.

This change merges PR #253.

Change-Id: I8907f68ea42065f3081f74f60752c5e6e93a93c5