Remove legacy mechanism for fallback SNMP scan functions
The legacy check API offers a mechanism to classify SNMP sections into
two categories: sections which are first scanned for and fallback
sections which are only considered if the former are not found. This
was used to prevent duplicate interface services. With the migration
of most interface checks to the new API, this mechanism has become
unnecessary.

CMK-5631

Change-Id: Id4d2c1d3087ace2c164521a3d03852df3e40d1c2