Always store user session information
Instead of just working with the user sessions when the "single
sessions" feature is enabled, the feature is now always enabled.

The intend is to use invalidate user login sessions not only on client
side (by removing a cookie), but also invalidating the session on the
server side to fully terminate the login session.

CMK-5080

Change-Id: I5023722624a3cdd9d18971d51e0fb60270b1e3de