Use emblems for the WATO snapins
The Quick access and WATO snapins now use emblems
as well. To achieve this the html.render_icon_button
and html.icon_button functions now have an optional
argument for the emblem.

Furthermore, the titles "WATO - Configuration" and
"WATO - Quickaccess" are replaced by "Setup" and
"Quick Setup". This is makes it clearer that they
belong to the "Setup" menu.

CMK-5485

Change-Id: I67a150cf07293c0051cca23200361fd5a150bdf0