Fix and clean cascadingdropdown choice and text render
CascadingDropdowns declare the identification of their value as the value
itself. When a choices can be configured over a ValueSpec, the
identification should be a string and the configured value a Tuple[str,
Any], yet mypy wont agree and wants Tuple[Ident, Any].(To be resolved)

The weak distinction between identification/value and configured value
allowed an abuse in Werk 4477, to set default values on the choices as a
configuration for the Timerange valuespec.

In this commit. Types for identification and value, are marked. For
non-disruptive compatibility on user set graph timeranges(Werk 4477), a
value_to_text is added to the Timerange.

CMK-5668

Change-Id: I74efb5315a8d77debb6db25698e1dd4334c016f6