Add login session ID to auth cookie
Previously we used a separate session cookie named "session_[site]" to
track the user login session ID in case the "single user session"
feature was active.

After we are now always storing and tracking the user sessions, there
was no point in separating both cookies. The session ID has now been
added to the auth cookie.

This change makes the pre 2.0 cookies incompatible with the new ones.
It means that all users will have to login into the GUI after the update
to Checkmk 2.0.

CMK-5499

Change-Id: I55f01ac6f36e78bad2b08f65d1073a62f1537847