Use sha256 for signing auth cookie values
Using md5 at this point is not a security problem, but it can be
considered bad practice.

In the past we have put this change on hold because we did not want to
break with the compatibility of cookies of old versions. Now that we
just changed the format incompatibly anyway, we can now also change the
hashing algorithms without additional consequences.

Change-Id: If14ad7b8a8052eb17ac9d7c0b81560d42111a30c