11345 FIX Use correct periods for AWS
The Special Agent for AWS uses the passed time since midnight for the cache interval for some sections.
However AWS does only allow specific values for the periods. This is now ensured with this werk.

Background information:
AWS expects "period" to be an integer and a multiple of 60 for metrics.
For metrics with high resolution, also the values 1, 5, 10, 30 and 60 are allowed.
See: https://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_MetricStat.html

Change-Id: I5f34544216f82a1c563420a940395fe3ff618c79