11496 User access times are now recorded and displayed by default
The feature "Save last access times of users" of the GUI, which was
disabled by default, is now always enabled. The time of the last access
is stored for each user.

The time of the last activity is shown on the users page for all admin
users by default. In case you want to hide this information, you can
remove the permission "Show last user activity" from your users.

CMK-5638

Change-Id: Ife51d6427ffb30a1677efeaeb6ec4971a257fc93