Check-API: prepare legacy conversion for stricter Result class
Bypass the validation of the Result class:
Legacy plugins may relie on the fact that once a newline
as been in the output, *all* following ouput is sent to
the details. That means we have to create Results with
details only, which will be prohibited by the Result class.

CMK-5431

Change-Id: I0d7a044d5c9d34c8e0cd43b56b602f8be9442200