11497 FIX Notification spooler: Fix potential cluster issues when trying to stop mknotifyd
When running a failover cluster (like on the Checkmk appliance) it is important
to stop all services in a reliable way before the cluster data volume can be
unmounted and moved over to the other nodes.

Previously the mknotifyd init script (which is used by omd stop) was not
waiting for the mknotifyd to terminate after signaling a stop to the process.
In rare cases this could lead to failover failures, when the mknotifyd takes
too long to stop.

Change-Id: Ib0a4498e4ca03cac21e4d5360d06c381477f493b