Show all builtin icons for a theme
Previously, the icon selector only displayed icons that were placed
in the directory of a theme. Since icons have an overriding mechanism
(i.e. all icons are placed in the facelift theme and only icons that
have to be adapted placed in other themes) not all builtin icons were
shown for the dark theme. This is fixed.

Change-Id: I09dba20bab394b54536d4e66c11e7169e05131fb