Fix regression in Werk 11392
According to pgrep(1):

The process name used for matching is limited to the 15 characters present
in the output of /proc/pid/stat. Use the -f option to match against the
complete command line, /proc/pid/cmdline.

Thus we can only match after pacemaker-contr, we can use -f as then it
matches to the entire command, which could be other command using crmd or
pacemaker-controld as arguments. We need to enforce a match on the process
name.

Change-Id: I7b4078142d48666a75d5e6e6fbd35f530abdfe64