Views: preserve filter popup state during page reloads
There are several situations where the filter form should be
shown on initial page rendering.

a) In case the user toggled the popup in the frontend and reloaded the
page, always enforce that state.

b) Show in case the view is a "mustsearch" view (User needs to submit
the filter form before data is shown).

c) Show after submitting the filter form. The user probably wants to
update the filters after first filtering.

CMK-5567

Change-Id: Ie97a0f3e70b765464dcfc99d585c71dfa2f11aa3