Make fetcher / checker respect --debug argument again
When debugging on the command line the --debug argument is often used to
make the command terminate with the traceback. Like for example: cmk
-nvv --debug myhost

In this situation we explicitly want to disable some exception handlings
which normally transport errors to the check results and really want to
terminate.

The recent fetcher / checker refactorings introduced some exception
handling which was not caring about the debug flag, preventing this kind
of debugging.

Change-Id: I0342d8b81a392a7b7d7c57d67242273e0d99dbf0