Service status border
For the moment the status border to service is rendered from the
single_value subplot, because I could insert that info in the plot
spec. Border is a separate subplot/component that can be hooked to the
figure base, yet the render information originates from the single value
recipe out of the python query. The component flexibility will then be
reused for other figure or a general refactor on the dashlet recipe
structure on future commits

- Display metric with unit scaling value

CMK-5743

Change-Id: I0dcba15b451333305404fbeec9d16d0dc9880742