Extend valuespecs to support transformation of given values
The valuespec classes have now been extended by a "transform_value" call
which task is to just execute the transformations of the encapsulated
Transform valuespects on the given data structure and return the
transformed data structure.

This will be used to migrate the values of existing rules during
Checkmk updates in a general way.

Change-Id: I704a81ad437975f1c42918d8a96a63c15f1a0333