11116 FIX job: Failed jobs are not OK just because a new one is running
The job plugin determines its state from the age of the last job
and its exit code.
Previously, a job was considered {OK} regardless of its exit code
whenever a job was running at the time of checking. This way a long
running job could appear to be {OK} most of the time, even if it
was constantly failing.

Change-Id: Iaa574e3631be678d74d13e886e31a321cbb1191c