ps: Cleanup & fix check parameters
Fixes a bug in the transformation of check parameters for the ps-checks
related to the fact that the Parameters object passed to check functions
by the new API is immutable. Also fixes a bug in the corresponding
transformation implemented in cmk/gui.

Also cleans up the check parameters for the check domino_tasks (which uses
the ps-functionality).

CMK-5806

Change-Id: I503565cb468d87ceed45d6fd427dfb6c2eb1e8ce