uses other values for folder usage
DELL-STORAGE-SC-MIB::scDiskFolderSUTotalSpace2 is the available space
for the folder
DELL-STORAGE-SC-MIB::scDiskFolderSUFreeSpace is the difference of
configured space and available space

Both values (available space and configured space) are configured by
the admin and do not change with the usage of the folder.

DELL-STORAGE-SC-MIB::scDiskFolderSUAllocSpace is the space allocated to
the folder (i.e. the total space)
DELL-STORAGE-SC-MIB::scDiskFolderSUUsedSpace2 is the space used by the
data stored in the folder

Therefor it seems more consistent to use the difference of the later two
values to compute the free space available for data in the folder.