dell_compellent_folder: Use correct MIBs
Currently, the following MIBs are used:
* DELL-STORAGE-SC-MIB::scDiskFolderSUTotalSpace2: available space
for the folder
* DELL-STORAGE-SC-MIB::scDiskFolderSUFreeSpace: difference between the
configured space and available space

Both values (available space and configured space) are configured by
the admin and do not change with the usage of the folder. Instead, the
follwoing MIBs should be used:

* DELL-STORAGE-SC-MIB::scDiskFolderSUAllocSpace: space allocated to the
folder (i.e. the total space)
* DELL-STORAGE-SC-MIB::scDiskFolderSUUsedSpace2: space used by the data
stored in the folder

The available space is given by the difference of the two.

This change also makes the check use the standard check parameters for
filesystem checks.

This merges PR #261, https://github.com/tribe29/checkmk/pull/261.

Change-Id: If44fe823e3a806f0c4b1e79b696064a21ed44dab