Introduce separate config setting for checker pool size
The fetcher and checker mode is scaling in a different way than the
previous Checkmk helper mode. This makes it necessary to tune the
pool sizes differently.

Instead of basing on the previously existing "Maximum concurrent Checkmk
checks" setting that controlled the number of "Checkmk helpers" in
previous Checkmk sites, we use a dedicated setting "Maximum concurrent
Checkmk checkers" for configuring the number of "Checkmk checkers".

All sites will start with a number of 4 checker processes. As these
processes are CPU bound, this should be a good start for the most
environments.

To tune the Checkmk check processing, you need to check both, the
Checkmk fetcher usage and the Checkmk checker usage and tune the pool to
a size which is able to handle your configuration and environment.

CMK-5741

Change-Id: I5da274fde8f100c3fbd95f3655a3ac6a67ea7f0f