Create search index when setting up a new site and during cmk-update-config
When setting up a new site, the search index needs to be created
initially, such that it can be used immediately. Furthermore, we
also re-build the index during cmk-update-config. This is also a
preparation for splitting the index into multiple smaller sub-
indices, which will allow us to do partial updates.

CMK-5866

Change-Id: I281434c317781fd86819ab650205c81e2a16bce7