Emblems can be selected in IconSelector now
It is now possible to select an emblem in an IconSelector if
with_emblem is set to True. To make this possible the icon
rendering methods in htmllib.py now accept either a str
which uses a plain icon or a TypedDict with an icon and
an emblem. The various call sites are adapted as well.
Right now the selector with an emblem is relatively simple
and will be changed in a later commit.

CMK-5775

Change-Id: I9b279f146facea8db28cdfda6f0cfcb2cc44b1ce