[CMK-5399] - add CPU tracking to fetcher
- track cpu usage for all fetchers
- deliver results of tracking as a log message(to be fixed in next CL) to Microcore
- add context manager API to cpu_tracking module
- unit test
- move cpu_tracking to cmk.utils to avoid layer violation warning
- introduce cpu_tracking.reset

This is first part of the ticket: gathering data.

Some details of implementation
Unfortunately current codebase requires that we have some kind of global data(decorator cpu_tracking.track) to track cpu usage. This approach makes impossible the transfer of the context of cpu tracking outside the module(like in cmk...profile).
We have chosen to use 'reset' together with starting of the gathering cpu info. This API(pair reset/start) may be changed in the future to be more elegant.

Change-Id: Ia0c3f3ac34372a2910f5b009efc54f8081c67dbe