Improve background job / automation debugging
The GUI log level now affects the verbosity of the output produced by
automation commands.

To get more diagnostic information on e.g. the service discovery page,
you you can now increase the log level for automations in the GUI
logging configuration which will a) give you more output of the GUI
mechanic and b) of the actual communication with the target device.

Previously the automations always used either "normal" output or
"verbose" output in case of the service discovery automation.

Change-Id: I198e017581962a8ab980f34fc5b015b49d95a9ff