Make method makeuri_contextless of cmk.gui.htmllib.html a function
The method cmk.gui.htmllib.html.makeuri_contextless is, among others,
used for retrieving the URLs of hosts and folders. This is also done
when updating the search index for the Setup menu after a change. The
REST API can make changes and thus trigger this step, however, it has no
access to the global html object in cmk.gui.globals. Hence, the current
implementation of the search index breaks the REST API. To resolve this,
we make makeuri_contextless a function that can be accessessed by the
API.

CMK-5885

Change-Id: Ibc2e8ec86271e418ddd8d4552f3f672e8ffd5f80