Fix slow service discovery page of SNMP hosts
Since 6c2961ce88e the SNMP fetcher cache was completely disabled when
fetching for service discovery. This is totally fine for all call sites,
except one: The service discovery page calls the automation
"try-inventory" with the "@noscan" option, e.g. on initial page
rendering.

In this situation all available caches should be used to make the page
render as fast as possible. This exception is now expressed in a special
fetcher mode called "CACHED_DISCOVERY".

CMK-5908

Change-Id: Id97c042006e3c25a388258a072f01d0e7701aefc