11373 FIX Made titles of options in host/folder permissions less uncertain
There have always been additional option when assigning contact groups to
WATO folders or hosts. The effect of this options meanwhile has not always
been absolute transparent depending if the user is in folder or in host
properties. Especially the outcome of option "Always add host contact groups
also to its services" has not been clear.

The titles do now explain better, what a particular option really does. This
is just a clarification of the current, actual behaviour and no change of
the behaviour itself. Therefore there is no need for manual interaction.

Change-Id: I99d4ad9395f694058b7a148e627b58d9a157020e