Improve GUI tests by mocking livestatus
Some unit tests accidentally tried to execute livestatus queries:

* Report rendering tests render views, which tried to get data from livestatus
* Rendering and processing filters in tests

We now use the generic livestatus mock to get get real data to show.
Only the WATO folder filter needed to be mocked, because it issues a
query with Stats header that is not understood by the livestatus mock at
the moment.

Bringing down these tests from 10 sec to 0.x sec.

Change-Id: Icd7970e08cd269e86622c8bf5dfd7ff8b0dd6459