Refactor to isolate title infos
The Cascadingdropdown has "plain" and "add_title_infos". Which means the
second option implies the first. The plain title is always there, and an
empty selection is equivalent to choosing plain, also on execution.

The title_format entry should only be a List choice to pick include:
- Custom title(Still to be implemented)
- Default title
- host name
- host alias
- service description
- metric name(placed not used yet)

In this way the user picks what is to be display, every selection is
concatenated.

There is the html render restriction to only allow this for template
graphs. Certainly because only there it correctly has the info on
host/service. It should be relaxed in the future.

CMK-5869

Change-Id: Ibb44fafd74dad8e553b7691911eef974b74fc275