11402 Agent Bakery: New "Bake and sign" action
It is now possible to bake and sign all agents in one step. You can simply do this by pressing
the corresponding button at the agent bakery. It will be available as soon as at least one
signature key has been created or imported.

When invoking "Bake and sign agents", you will be prompted for selecting a signature key and
entering its passphrase. After that, every agent package will get baked and signed immediately
after baking.

When using "Bake and sign", it is assured that there are no unsigned packages visible for
download at any time, avoiding that it could be seen accidentally by a calling agent updater,
leading to a "The deployment server provides an agent but that is not signed." error.

The "Bake and sign" action is also available via REST API, please refer to the documentation
for details.

CMK-4680

Change-Id: I47706ea843b1c7f0e53561d6e7e5d87ea39d2930