Performance graph dashlet title on header top
Dashlets for Performance Graphs and Custom graphs had those names a dashlet
title. The title of the graph was rendered inside of it, by the graph
system. Now the graph title is extracted and displayed as dashlet title.

Extend option dashlet show title

The checkbox boolean turns in dropdown. No transform necessary because, the
old options keep their type choice. This dropdown is similar in style to
the show_title for the performance graphs.

The show_title and title_format, need to be extracted out of
graph_rendering_options, yet they are still needed there, as the graph
title can for the moment only be rendered there. For the moment the
show_title option is removed from the render graph options on dashlets and
given the false default, thus title wont appear repeated in graph.

This behavior should simplify once all dashlets are rendered over D3.js

CMK-5878

Change-Id: Ie10a536b524efd15ef96f8d39c5fc6d717f9afc1