Refactor graph dashlet, graph identification resolved at init
Move the pnp graph render options to init. This way I set the title format
in old standard for inside the graph, and make the default also available
for future on dashlet title.

The graph inside a dashlet has always the same identification, it does not
need to be computed on every js_reload.

- Do the same for custom graphs dashlet

CMK-5870

Change-Id: I59f04fb932aee3666e4b9124f56cb5b8e720e35e