cpu_tracking: Pass stats with payload #1
The stats correspond to the payload, indeed we measure the time needed
to obtain the payload.  It is therefore meaningful to group them.

This further lets us group serialization and deserialization of the
timings in a way that is consistent with the rest of the fetcher/checker
serialization.

Now, the stats are still passed in a different message than the payload.

CMK-5909

Change-Id: I6012b2f33f554913328b7c85543fc16eed88483c