Style Gauge dashlet
Title is directly rendered over the with render_title component. Margin is
auto adjusted to allow space for it. Figure and plot size are
correspondingly changed.

Handle all placing of elements within the g element, no more per component
translation.

Current value is raised in position. It for the moment collides with the
pointer, yet pointer will be removed in future iterations to match the UX
design.

Include the dashlet border status component.

CMK-5870

Change-Id: I237f5d01749e8d68f03a04e6225d173ea3b82070