Improve performance of discovery page actions
When enabling/disabling services, the set-autochecks automation is often
called to store the just changed autochecks.

In the past the checks had to be loaded by this automation call, just to
be able to compute the service description (which was always available
in the calling GUI code, but not handed over to the automation).

To solve this, the GUI now hands over the previously calculated service
description which makes the automation call much faster.

CMK-5965

Change-Id: If9ec2a88a0fecd0bbb60f60d3301683660b877a7